requirements.txt
customtkinter==5.2.2
python-docx==1.2.0
pyinstaller==6.19.0
packaging==25.0
